% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vash.R
\name{est_mixprop_mode}
\alias{est_mixprop_mode}
\title{Estimate mixture proportions and mode of the unimodal inverse-gaama mixture variance prior.}
\usage{
est_mixprop_mode(sehat, g, prior, df, unimodal, singlecomp, estpriormode,
  maxiter = 5000)
}
\arguments{
\item{sehat}{n vector of standard errors of observations}

\item{g}{the initial prior distribution for variances}

\item{prior}{numeric vector indicating Dirichlet prior on mixture proportions}

\item{df}{appropriate degrees of freedom for chi-square distribution of sehat^2}

\item{unimodal}{put unimodal constraint on the prior distribution of variances ("variance") or precisions ("precision")}

\item{singlecomp}{logical, indicating whether to use a single inverse-gamma distribution as the prior distribution for the variances}

\item{estpriormode}{logical, indicating whether to estimate the mode of the unimodal prior}

\item{maxiter}{maximum number of iterations of the EM algorithm}
}
\value{
A list, including the final loglikelihood, the fitted prior g, number of iterations and a flag to indicate convergence.
}
\description{
Estimate mixture proportions and mode of the unimodal inverse-gaama mixture variance prior.
}
\examples{
fitted.prior = est_mixprop_mode(sehat=abs(rnorm(100)),g=igmix(c(.5,.5),c(1,3),c(1,3)),
prior=c(1,1),df=10,unimodal="variance",singlecomp=FALSE,estpriormode=TRUE)

}

