% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vash.R
\name{loglike}
\alias{loglike}
\title{Compute log-likelihood of the variance model with mixture inverse-gamma prior}
\usage{
loglike(sehat, df, pi, alpha, beta)
}
\arguments{
\item{sehat}{a p vector of observed standard errors}

\item{df}{the degree of freedom}

\item{pi}{a k vector, the mixture proportions of the k component inverse-gamma mixture prior}

\item{alpha}{a k vector, the shape parameters of the k component inverse-gamma mixture prior}

\item{beta}{a k vector, the rate parameters of the k component inverse-gamma mixture prior}
}
\value{
log-likehihood
}
\description{
Suppose we observe standard errors sehat, where \eqn{sehat^2~s^2 \times \chi^2_{df}/df}, and $s^2$ comes from an inverse-gamma mixture prior. This function computes the log-likelihood of the model.
}
\examples{
sehat = abs(rnorm(10))
loglike(sehat, df=10, pi=c(0.1,0.3,0.5), alpha=c(5,8,10), beta=c(2,4,9))

}

