% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vash.R
\name{mod_t_test}
\alias{mod_t_test}
\title{Moderated t-test with standard errors moderated by mixture prior}
\usage{
mod_t_test(betahat, se, pi, df)
}
\arguments{
\item{betahat}{a p vector of observed effect sizes}

\item{se}{a p*k matrix of moderated standard errors}

\item{pi}{a p*k matrix of mixture proportions, row sums are 1.}

\item{df}{a p*k matrix of degrees of freedom}
}
\value{
A p vector of p-values testing if the effect sizes are 0.
}
\description{
Moderated t-test with standard errors moderated by mixture prior.
}
\examples{
# p=10, k=5
betahat = rnorm(10)
se = matrix(abs(rnorm(10*5)), ncol=5)
pi = matrix(rep(0.2,10*5), ncol=5)
df = matrix(rep(11:15,each=10), ncol=5)
mod_t_test(betahat,se,pi,df)

}

