% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vash}
\alias{vash}
\title{Main Variance Adaptive SHrinkage function}
\usage{
vash(sehat, df, betahat = NULL, randomstart = FALSE, singlecomp = FALSE,
  unimodal = c("variance", "precision"), prior = NULL, g = NULL,
  maxiter = 5000)
}
\arguments{
\item{sehat,}{a p vector of standard errors}

\item{df:}{appropriate degrees of freedom for (chi-square) distribution of sehat}

\item{betahat:}{a p vector of estimates}

\item{randomstart:}{bool, indicating whether to initialize EM randomly. If FALSE, then initializes to prior mean (for EM algorithm) or prior (for VBEM)}

\item{singlecomp:}{bool, indicating whether to use a single inverse-gamma distribution as the prior distribution for the variances}

\item{unimodal:}{unimodal constraint for the prior distribution of the variances ("variance") or the precisions ("precision")}

\item{prior:}{string, or numeric vector indicating Dirichlet prior on mixture proportions (defaults to "uniform", or 1,1...,1; also can be "nullbiased" 1,1/k-1,...,1/k-1 to put more weight on first component)}

\item{g:}{the prior distribution for beta (usually estimated from the data; this is used primarily in simulated data to do computations with the "true" g)}

\item{maxiter:}{maximum number of iterations of the EM algorithm}
}
\value{
a list with elements fitted.g is fitted mixture
}
\description{
Takes vectors of estimates (betahat) and their standard errors (sehat), and applies
shrinkage to them, using Empirical Bayes methods, to compute shrunk estimates for beta.
}
\details{


See readme for more details
}
\examples{
se = rigamma(100,1,1)
sehat = se*rchisq(100,5)/5
se.vash = vash(sehat,5)
plot(sehat,se.vash$PosteriorMean,xlim=c(0,10),ylim=c(0,10))
}

