% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vash.R
\name{est_singlecomp_mm}
\alias{est_singlecomp_mm}
\title{Estimate the single component inverse-gamma prior by matching the moments}
\usage{
est_singlecomp_mm(sehat, df)
}
\arguments{
\item{sehat}{n vector of standard errors of observations}

\item{df}{degrees of freedom for chi-square distribution of sehat^2}
}
\value{
The shape and rate of the inverse-gamma prior
}

