% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vash.R
\name{mod_t_test}
\alias{mod_t_test}
\title{Moderated t-test with standard errors moderated by mixture prior}
\usage{
mod_t_test(betahat, se, pi, df)
}
\arguments{
\item{betahat}{a p vector of observed effect sizes}

\item{se}{a p*k matrix of moderated standard errors}

\item{pi}{a p*k matrix of mixture proportions}

\item{df}{a p*k matrix of degrees of freedom}
}
\value{
A p vector of p-values
}

